BeautifulSoup/requirements.txt
fastapi
uvicorn
beautifulsoup4
lxml
python-dotenv
requests
openai
streamlit
